refactor(shared): move component to sub dir
To skip route creation for this component
