+++ src/playground/layout/shared/components/dynamic.component.ts
